fix: labels/annot logs and values.
